+++ lib/src/command/sub_command/generate_bloc.dart
@@ -10,13 +10,21 @@ class GenerateBlocSubCommand extends CommandBase {
         abbr: 'n', negatable: false, help: 'no create file test'
         //Add in future configured the release android sign
         );
+    argParser.addFlag('flutter_bloc',
+        abbr: 'f', negatable: true, help: 'using flutter_bloc package'
+        //Add in future configured the release android sign
+        );
+    argParser.addFlag('mobx',
+        abbr: 'm', negatable: true, help: 'using mobx package'
+        //Add in future configured the release android sign
+        );
   }
 
   void run() {
     if (argResults.rest.isEmpty) {
       throw UsageException("value not passed for a module command", usage);
     } else {
-      Generate.bloc(argResults.rest.first, !argResults["notest"]);
+      Generate.bloc(argResults.rest.first, !argResults["notest"], argResults["flutter_bloc"], argResults["mobx"]);
     }
   }
 }